🐛 Bind methods accessed by data
Consider a component like this:

```typescript
@Component
class MyComponent {
  foo = 'foo'
  options = {
    handler = this.method,
  }

  method() {
    return this.foo
  }
}
```

When we call `options.handler()`, we would expect to get back `foo`,
but this currently doesn't work, because the methods aren't correctly
bound in the data builder function.

This change actively binds the functions to the Vue component instance,
so that the above example works as expected.